Full rewrite of the account edition.
The account edition is now managed by an OverlaySheet, similar to the
one used for account creation. However, the account's name and password
cannot be modified this way.

The account name may be changed in the future, but this will require
special Backend code (it will be a rename operation instead of juste
updating a field).

The password modification will be managed entirely differently, because
it will use the password history.